requirements.txt

typing-extensions
pydantic
html2text
fastapi
uvicorn
requests
beautifulsoup4
aiohttp
markdownify
python-dotenv
Pillow
google-generativeai